# docs/swagger.yaml
definitions:
...
  utils.Result:
    properties:
      code:
        description: 状态码
        example: 100
        type: integer
      data:
        description: 数据
        type: object
      msg:
        description: 提示
        example: 密码错误
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  description: 博客项目 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Swagger
  version: "1.0"
paths:
...
  /api/v1/sys_setting/items:
    get:
      consumes:
      - application/json
      parameters:
      - description: 设置名称
        in: query
        name: name
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          description: ""
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取设置条目
      tags:
      - 系统设置
...
swagger: "2.0"